challenges/web/cookie-monster/src/requirements.txt
pyjwt
flask